kafka streams

+++ project3/src/main/java/project3/data/Item.java
@@ -14,11 +14,11 @@ public class Item implements Serializable {
         int id;
 
         private String name;
-        private float price;
+        private int price;
 
         public Item() {}
 
-        public Item(int id, String name, float price) {
+        public Item(int id, String name, int price) {
             this.id = id;
             this.name = name;
             this.price = price;
@@ -40,15 +40,15 @@ public class Item implements Serializable {
             this.name = name;
         }
 
-        public float getPrice() {
+        public int getPrice() {
             return price;
         }
 
-        public void setPrice(float price) {
+        public void setPrice(int price) {
             this.price = price;
         }
 
-        public void update(String name, float price) {
+        public void update(String name, int price) {
             this.name = name;
             this.price = price;
         }
